Moved FilterBuilder file
Uploaded file to wrong directory.
